+++ Branch-TestBed/Branch-SDK-Tests/BNCPasteboardTests.m
@@ -31,43 +31,33 @@
 
 - (void)addStringToPasteboard {
 #if !TARGET_OS_TV
-    if (@available(iOS 10.0, *)) {
-        [UIPasteboard.generalPasteboard setString:self.testString];
-    }
+    [UIPasteboard.generalPasteboard setString:self.testString];
 #endif
 }
 
 - (void)addBranchURLToPasteboard {
 #if !TARGET_OS_TV
-    if (@available(iOS 10.0, *)) {
-        [UIPasteboard.generalPasteboard setURL:self.testBranchURL];
-    }
+    [UIPasteboard.generalPasteboard setURL:self.testBranchURL];
 #endif
 }
 
 - (void)addNonBranchURLToPasteboard {
 #if !TARGET_OS_TV
-    if (@available(iOS 10.0, *)) {
-        [UIPasteboard.generalPasteboard setURL:[NSURL URLWithString:@"https://www.apple.com"]];
-    }
+    [UIPasteboard.generalPasteboard setURL:[NSURL URLWithString:@"https://www.apple.com"]];
 #endif
 }
 
 - (void)clearPasteboard {
 #if !TARGET_OS_TV
-    if (@available(iOS 10.0, *)) {
-        // cannot delete items from the pasteboard, but we can put something else on there
-        [[UIPasteboard generalPasteboard] setString:@""];
-    }
+    // cannot delete items from the pasteboard, but we can put something else on there
+    [[UIPasteboard generalPasteboard] setString:@""];
 #endif
 }
 
 - (NSString *)getStringFromClipboard {
     NSString *string = nil;
 #if !TARGET_OS_TV
-    if (@available(iOS 10.0, *)) {
-        string = [UIPasteboard.generalPasteboard string];
-    }
+    string = [UIPasteboard.generalPasteboard string];
 #endif
     return string;
 }
@@ -75,9 +65,7 @@
 - (NSURL *)getURLFromPasteboard {
     NSURL *url = nil;
 #if !TARGET_OS_TV
-    if (@available(iOS 10.0, *)) {
-        url = [UIPasteboard.generalPasteboard URL];
-    }
+    url = [UIPasteboard.generalPasteboard URL];
 #endif
     return url;
 }
@@ -157,7 +145,7 @@
 // This test fails intermittently when executed with other tests - depending upon the order in which its executed
 - (void) testPassPasteControl {
 #if !TARGET_OS_TV
-    if (@available(iOS 16.0, *)) {
+    if (@available(iOS 16.0, macCatalyst 16.0, *)) {
         
         long long timeStamp = ([[NSDate date] timeIntervalSince1970] - 5*60)*1000; // 5 minute earlier timestamp
         NSString *urlString = [NSString stringWithFormat:@"https://bnctestbed-alternate.app.link/9R7MbTmnRtb?__branch_flow_type=viewapp&__branch_flow_id=1105940563590163783&__branch_mobile_deepview_type=1&nl_opt_in=1&_cpts=%lld", timeStamp];