+++ BranchSDK/BNCDeviceInfo.h
@@ -53,7 +53,6 @@
 @property (nonatomic, copy, readwrite) NSNumber *screenWidth;
 @property (nonatomic, copy, readwrite) NSNumber *screenHeight;
 @property (nonatomic, copy, readwrite) NSNumber *screenScale;
-@property (nonatomic, copy, readwrite) NSString *carrierName;
 @property (nonatomic, copy, readwrite) NSString *locale;
 @property (nonatomic, copy, readwrite) NSString *country; //!< The iso2 Country name (us, in,etc).
 @property (nonatomic, copy, readwrite) NSString *language; //!< The iso2 language code (en, ml).