+++ BranchSDK/BranchOpenRequest.m
@@ -95,7 +95,7 @@
         [self safeSetValue:partnerParameters forKey:BRANCH_REQUEST_KEY_PARTNER_PARAMETERS onDict:params];
     }
         
-    if (@available(iOS 16.0, *)) {
+    if (@available(iOS 16.0, macCatalyst 16.0, *)) {
         NSString *localURLString = [[BNCPreferenceHelper sharedInstance] localUrl];
         if(localURLString){
             NSURL *localURL = [[NSURL alloc] initWithString:localURLString];
@@ -290,7 +290,7 @@ typedef NS_ENUM(NSInteger, BNCUpdateState) {
         NSNumber *invokeRegister = (NSNumber *)data[BRANCH_RESPONSE_KEY_INVOKE_REGISTER_APP];
         preferenceHelper.invokeRegisterApp = invokeRegister.boolValue;
         if (invokeRegister.boolValue && self.isInstall) {
-            if (@available(iOS 16.1, *)){
+            if (@available(iOS 16.1, macCatalyst 16.1, *)){
                 NSString *defaultCoarseConValue = [[BNCSKAdNetwork sharedInstance] getCoarseConversionValueFromDataResponse:@{}];
                 [[BNCSKAdNetwork sharedInstance] updatePostbackConversionValue:0 coarseValue:defaultCoarseConValue
                     lockWindow:NO completionHandler:^(NSError * _Nullable error) {
@@ -300,7 +300,7 @@ typedef NS_ENUM(NSInteger, BNCUpdateState) {
                         BNCLogDebug([NSString stringWithFormat:@"Update conversion value was successful for INSTALL Event"]);
                     }
                 }];
-            } else if (@available(iOS 15.4, *)){
+            } else if (@available(iOS 15.4, macCatalyst 15.4, *)){
                 [[BNCSKAdNetwork sharedInstance] updatePostbackConversionValue:0 completionHandler:^(NSError * _Nullable error) {
                     if (error) {
                         BNCLogError([NSString stringWithFormat:@"Update conversion value failed with error - %@", [error description]]);
@@ -322,7 +322,7 @@ typedef NS_ENUM(NSInteger, BNCUpdateState) {
         NSNumber *conversionValue = (NSNumber *)data[BRANCH_RESPONSE_KEY_UPDATE_CONVERSION_VALUE];
         // Regardless of SKAN opted-in in dashboard, we always get conversionValue, so adding check to find out if install/open response had "invoke_register_app" true
         if (conversionValue && preferenceHelper.invokeRegisterApp ) {
-            if (@available(iOS 16.1, *)){
+            if (@available(iOS 16.1, macCatalyst 16.1, *)){
                 NSString* coarseConversionValue = [[BNCSKAdNetwork sharedInstance] getCoarseConversionValueFromDataResponse:data] ;
                 BOOL lockWin = [[BNCSKAdNetwork sharedInstance] getLockedStatusFromDataResponse:data];
                 BOOL shouldCallUpdatePostback = [[BNCSKAdNetwork sharedInstance] shouldCallPostbackForDataResponse:data];
@@ -338,7 +338,7 @@ typedef NS_ENUM(NSInteger, BNCUpdateState) {
                         }
                     }];
                 }
-            } else if (@available(iOS 15.4, *)) {
+            } else if (@available(iOS 15.4, macCatalyst 15.4, *)) {
                 [[BNCSKAdNetwork sharedInstance] updatePostbackConversionValue:conversionValue.intValue completionHandler: ^(NSError *error){
                     if (error) {
                         BNCLogError([NSString stringWithFormat:@"Update conversion value failed with error - %@", [error description]]);