+++ BranchSDK/BranchShareLink.h
@@ -93,7 +93,7 @@ Presents a UIActivityViewController that shares the Branch link.
 @property (nonatomic, strong, nullable) NSURL *placeholderURL;
 
 // iOS 13+ : LinkPresentation metadata for the preview header.
-@property (nonatomic, strong, nullable) LPLinkMetadata *lpMetaData API_AVAILABLE(ios(13.0));
+@property (nonatomic, strong, nullable) LPLinkMetadata *lpMetaData API_AVAILABLE(ios(13.0), macCatalyst(13.1));
 
 ///Share text for the item.  This is not the text in the iOS 13+ preview header.
 ///This text can be changed later when the `branchShareSheetWillShare:` delegate method is called.
@@ -135,7 +135,6 @@ Creates and attaches an LPLinkMetadata using the provided title and icon. This m
 @param icon             The image used for the share sheet preview icon.
 */
 
-- (void) addLPLinkMetadata:(NSString *_Nullable)title
-                      icon:(UIImage *_Nullable)icon API_AVAILABLE(ios(13.0));
+- (void) addLPLinkMetadata:(NSString *_Nullable)title icon:(UIImage *_Nullable)icon API_AVAILABLE(ios(13.0), macCatalyst(13.1));
 
 @end